+++ src/main/java/nextstep/security/authentication/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.authentication;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
