+++ src/main/java/nextstep/security/authentication/AuthenticationManager.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.authentication;
 
 public interface AuthenticationManager {
     Authentication authenticate(Authentication authentication);
