+++ src/main/java/nextstep/security/authentication/UsernamePasswordAuthenticationToken.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.authentication;
 
 public class UsernamePasswordAuthenticationToken implements Authentication {
     private final String principal;
