+++ src/main/java/nextstep/security/filter/BasicAuthenticationFilter.java
@@ -4,10 +4,10 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.security.Authentication;
-import nextstep.security.AuthenticationManager;
-import nextstep.security.SecurityContextHolder;
-import nextstep.security.UsernamePasswordAuthenticationToken;
+import nextstep.security.authentication.Authentication;
+import nextstep.security.authentication.AuthenticationManager;
+import nextstep.security.authentication.UsernamePasswordAuthenticationToken;
+import nextstep.security.context.SecurityContextHolder;
 import nextstep.security.util.Base64Convertor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
