+++ src/test/java/nextstep/security/authentication/ProviderManagerTest.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.authentication;
 
 import nextstep.security.fixture.TestAuthentication;
 import nextstep.security.fixture.TestAuthenticationProvider;
