+++ src/test/java/nextstep/security/authentication/UsernamePasswordAuthenticationTokenTest.java
@@ -1,4 +1,4 @@
-package nextstep.security;
+package nextstep.security.authentication;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
